refactor: extract interactive mode helpers from cli.py (#925)
Move rendering, file-watching, and session index helpers into a new
interactive.py module. cli.py re-exports all moved symbols under their
original private names so existing tests pass unchanged.

- New: src/copilot_usage/interactive.py (~160 lines)
  - print_version_header, render_session_list, draw_home, write_prompt
  - FileChangeHandler, FileChangeEventHandler, Stoppable protocols
  - start_observer, stop_observer, build_session_index
- cli.py keeps: Click commands, date parsing, _interactive_loop,
  _read_line_nonblocking, _show_session_by_index
- Updated architecture.md Components table
